fix: resolve all backend compilation errors from remote merge
- DataImportJob: remove duplicate ImportStatus enum, remove redundant
  accessor methods, use setters in factory method
- CustomRole/SystemSetting: use setter methods instead of direct field
  access to BaseEntity private fields
- OversightService: add missing repository injections (AssessmentRepository,
  SubjectRepository), qualify inner DTO classes (PerformanceResponse.SchoolPerformance etc),
  fix LiveClass entity method calls to use UUID-based lookups, fix repository
  method signatures
- ParentSelfController/ParentDTOs: add missing fields to AssessmentItem,
  CourseProgress, NotificationItem DTOs; fix unreadCount type
- LiveClassParticipantRepository: add existsByLiveClassIdAndUserIdAndIsDeletedFalse
- ParentDashboardServiceImpl: fix unreadCount cast
- pom.xml: remove incorrect -sourcepath compiler override

+++ backend/elmkusoma-core/src/main/java/tz/elmkusoma/parent/service/impl/ParentDashboardServiceImpl.java
@@ -589,7 +589,7 @@ public class ParentDashboardServiceImpl implements ParentDashboardService {
 
         return ParentNotificationResponse.builder()
                 .notifications(notifications)
-                .unreadCount((int) notifications.stream().filter(n -> !Boolean.TRUE.equals(n.getIsRead())).count())
+                .unreadCount(notifications.stream().filter(n -> !Boolean.TRUE.equals(n.getIsRead())).count())
                 .build();
     }
